+++ apps/meteor/app/gitlab/server/index.ts
@@ -1,1 +1,1 @@
-import '../lib/common';
+import './lib';
