+++ packages/core-typings/src/IUser.ts
@@ -93,7 +93,7 @@ export interface IUserServices {
 		serverURL: string;
 	};
 	dolphin?: {
-		Nickname?: string;
+		NickName?: string;
 	};
 }
